requirements.txt
python-mpd2
mutagen
toml
appdirs